Completed video plugin.
