+++ plugin_tutorial_2_video_cross_painter/gst-plugin/src/boilerplate.h
