+++ plugin_tutorial_2_video_cross_painter/main.cpp
